BioMonitor v0.0.1 BioMonitor.Router.Helpers

Module with named helpers generated from BioMonitor.Router.

Functions

path(data, path)

Generates the path information including any necessary prefix.

ph_path(conn_or_endpoint, action)

ph_path(conn_or_endpoint, action, params)

ph_url(conn_or_endpoint, action)

ph_url(conn_or_endpoint, action, params)

routine_path(conn_or_endpoint, action)

routine_path(conn_or_endpoint, action, params)

routine_path(conn_or_endpoint, action, id, params)

routine_reading_path(conn_or_endpoint, action, routine_id)

routine_reading_path(conn_or_endpoint, action, routine_id, params)

routine_reading_path(conn_or_endpoint, action, routine_id, id, params)

routine_reading_url(conn_or_endpoint, action, routine_id)

routine_reading_url(conn_or_endpoint, action, routine_id, params)

routine_reading_url(conn_or_endpoint, action, routine_id, id, params)

routine_routine_path(conn_or_endpoint, action, routine_id)

routine_routine_path(conn_or_endpoint, action, routine_id, params)

routine_routine_url(conn_or_endpoint, action, routine_id)

routine_routine_url(conn_or_endpoint, action, routine_id, params)

routine_url(conn_or_endpoint, action)

routine_url(conn_or_endpoint, action, params)

routine_url(conn_or_endpoint, action, id, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.