BioMonitor v0.0.1 BioMonitor.RoutineProcessing

The RoutineProcessing module has all the helper functions to manage the actions that need to be done on a routine.

Functions

add_value_to_ph_cal(new_value, ph_cal)

check_for_pump_trigger(routine, timestamp)

check_for_triggers(routine, start_timestamp)

fetch_reading(routine_id)

get_current_temp_target(routine, timestamp)

get_sensors_status()

is_offset_stable?(ph_cal)

process_reading(arg1, routine, temp_target)

register_error(message)

save_routine_sart_timestamp(routine)

send_ph_offset(target, offset)

start_ph_balance(ph, target_ph, tolerance)