BioMonitor v0.0.1 BioMonitor.SensorManager

Wrapper module around all serial communication with all sensors using a SerialMonitor instance.

Functions

arduino_gs_id()

Helper to expose arduino’s serial monitor identifier.

get_ph()

Fetch the ph value from the sensor.

get_ph_offset()

get_readings()

Fetchs all readings from the SerialMonitors and parse them.

pump_acid()

sets on the acid pump to drop.

pump_base()

sets on the base pump to drop.

pump_trigger(for_seconds)

pumps through the third bomb any substance that the operator has set up for a time interval.

push_acid()

pushes acid through the pump

push_base()

pushes base through the pump

send_and_read(sensor, command)

Sends a command for an specific sensor and reads the response. sensor should be one of the previously reigstered sensors.

example send_command(:temp, “getTemp”) returns:

* {:ok, result}
* {:error, message}

send_command(sensor, command)

Sends a command for an specific sensor. sensor should be one of the previously reigstered sensors.

example send_command(:temp, “getTemp”) returns:

* {:ok, result}
* {:error, message}

set_ph_offset(target, value, offset)

sets the offset of the ph sensor for calibration

start_sensors()

Adds all sensors specified in the config file to the SerialMonitor.