BioMonitor v0.0.1 BioMonitor.SerialMonitor

The SerialMonitor is in charge of the serial communication with the sensors

Functions

add_sensors(pid, sensors)

Adds the sensors whose information comes through the connected device

pid: The PID of the running process sensors: Key value pair, where the key is an atom representing the type of

     the sensor, and the value is the command to fetch the data

Examples

add_sensor [temp: “getTemp”, ph: “getPh”]

clean_serial(pid)

get_ports(pid)

Returns a list of available ports and the devices connected to them

pid: The PID of the running process

get_readings(pid)

Retrieves the readings for the registered sensors

pid: The PID of the running process

Examples

If :temp and :ph sensors are defined, then the returned value would be:

{:ok, [temp: “20.75”, ph: {:error, “Unrecognized Command”}]}

handle_call(msg, from, state)

Sends the command to to the device

init(args)

State structure: %{ serial_pid: The pid of the process that handles serial connectivity port: The port information to establish a connection with the device sensors: Key value pair, where the key is an atom representing the type of

       the sensor, and the value is the command to fetch the data

}

Examples

%{ serial_pid: 1234, port: %{

name: The name of the serial port where the device is connected
speed: The speed of the serial port

} sensors: [

%{temp => "getTemp"},
%{ph => "getPh"}

] }

send_and_read(pid, command)

send_command(pid, command)

Sends a command to the sensor.

pid: The PID of the running process

command: The String instruction to be sent to the sensor

set_port(pid, name, speed)

Sets the port where the device is connected

pid: The PID of the running process name: The name identifier of the port (use get_ports/0 to retrieve it) speed: The speed of the serial port

start_link(opts)